pyproject.toml
[tool.poetry]
name = "markdown-convert"
version = "0.1.0"
description = ""
authors = ["Tiago Rodrigues <tiagopbr.2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.12"
markitdown = {extras = ["all"], version = "^0.1.1"}
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
